buff sub quick fix (#196)
This is just a quick fix for the subtitle problem when moving the buff list. Now the subtitles will go back to its original position once the buff list is high enough. 

The threshold depends on the current hud scale and the result is far from perfect but it will prevent the subtitles to go higher than it needs to. Right now the subtitles will go as high as to just below the interaction circle when having the lowest hud scale and the buff list around 250-260.